+++ src/main/java/fr/bryanprolong/gatherthem/gatherthem_server/collections/exposition/dtos/CollectionDto.java
@@ -7,7 +7,7 @@ public class CollectionDto {
     private String type;
     private String name;
     private String description;
-    private int creationDate;
+    private long creationDate;
 
     public String getId() {
         return id;
@@ -41,11 +41,11 @@ public class CollectionDto {
         this.description = description;
     }
 
-    public int getCreationDate() {
+    public long getCreationDate() {
         return creationDate;
     }
 
-    public void setCreationDate(int creationDate) {
+    public void setCreationDate(long creationDate) {
         this.creationDate = creationDate;
     }
 }
